+++ src/log.c
@@ -16,30 +16,30 @@
 
 void	log_state_change(t_philo philo, char *state)
 {
-	long long	timestamp;
+	unsigned int	timestamp;
 
 	pthread_mutex_lock(&philo.simulation->print_lock);
 	timestamp = get_delta_time(philo.simulation->start_time);
-	printf("%lld %d is %s\n", timestamp, philo.id, state);
+	printf("%u %d is %s\n", timestamp, philo.id, state);
 	pthread_mutex_unlock(&philo.simulation->print_lock);
 }
 
 void	log_fork(t_philo philo)
 {
-	long long	timestamp;
+	unsigned int	timestamp;
 
 	pthread_mutex_lock(&philo.simulation->print_lock);
 	timestamp = get_delta_time(philo.simulation->start_time);
-	printf("%lld %d has taken a fork\n", timestamp, philo.id);
+	printf("%u %d has taken a fork\n", timestamp, philo.id);
 	pthread_mutex_unlock(&philo.simulation->print_lock);
 }
 
 void	log_death(t_philo philo)
 {
-	long long	timestamp;
+	unsigned int	timestamp;
 
 	pthread_mutex_lock(&philo.simulation->print_lock);
 	timestamp = get_delta_time(philo.simulation->start_time);
-	printf("%lld %d died\n", timestamp, philo.id);
+	printf("%u %d died\n", timestamp, philo.id);
 	pthread_mutex_unlock(&philo.simulation->print_lock);
 }
